Add resolve and heal infrastructure for 2pc
This commit adds the ability to resolve whether or not
2pc transactions have been committed or aborted and also
adds a heal function to resolve transactions that have
been prepared but not committed or rolled back.

This commit also removes the server id of the primary
key on the remote_txn table and adds another index.
This was done because the `remote_txn_persistent_record_exists`
should not rely on the server being contacted but
should rather just check for the existance of the id.
This makes the resolution safe to setups where
two frontend server definitions point to the same database.
While this may not be a properly configured setup, it's better
if the resolution process is robust to this case.