Refactor the planning of foreign modifies in FDW
This cleans up the code in the PlanForeignModify callback of the
foreign data wrapper and adds test output to validate that
proper SQL statements are generated.

A minor fix has been applied to the deparsing code to allow us to
deparse INSERT/UPDATE/DELETE statements for a relation that might not
be a foreign table. We need to be able to deparse statements for
regular relations since modifications on the remote data node will
happen on the main hypertable, and not directly on chunks.