Query hypertables on a per-server instead of a per-chunk basis
This optimization enhances the query when querying a hypertable
using the timescale fdw. Previously such queries created execution
nodes and queries on a per-chunk basis. This PR combines all the chunks
belonging to the same hypertable and server together so that only one
query and executor node are created per hypertable-server.

This is accomplished by first changing the chunk expansion code to
not do the table expansion for remote hypertables and instead simply
save the chunk oids in a metadata field in TimescaleDBPrivate.

Next, we intercept the set rel pathlist hook to create paths for the
hypertable-server nodes. This uses the new server chunk assignment
module to choose which chunks will be fetched from which servers. For
now, we have only one assignment strategy but later we can have multiple
strategies each creating it's own paths and having the planner choose
the cheapest path using standard methodologies.

Finally, during the plan creation phase we pass down the server chunk
assignment to the deparser so that it can add a `chunks_in` call to the
where clause. This tells the data node which chunks to use and is
necessary when we have replicated chunks so that two servers don't
return data for the same chunk. An alternative approach for deparsing
`chunks_in` could have been to add a clause to remote_exprs and have
the deparser include that in the WHERE clause that way. However, the
standard way to deparse whole-row expressions is using the ROW syntax
(to protect against different schema definitions on local and remote
nodes). However, 'chunks_in' requires a record, not row reference
so that approach was rejected as being too awkward. Some additional
deparsing changes were made to handle base relations that don't have
an associated foreign-table (i.e. the hypertable-server case).

This commit also changes the way fdw_private is treated on RelOptInfo.
Previously, this could have been NULL, a TimescaleDBPrivate object
or a TsFwdRelationInfo. This led to some bugs and awkwardness as
well as increased lack of type-safety. Now, this field could only
be NULL or TimescaleDBPrivate and the TsFwdRelationInfo object
is now an optional and type-safe member of TimescaleDBPrivate.