Make chunk placement consistent across servers
Chunks are placed across servers based on the ordinal of the slice in
the first space dimension, if such a dimension exists. For instance,
if a chunk belongs to the second slice in the space dimension, this
ordinal number will be used modulo the number of servers to find the
server to place the chunk on.

However, the ordinal is calculated based on the existing slices in the
dimension, and, because slices are created lazily, the ordinal of a
specific slice might vary until all slices are created in the space
dimension. This has the result that chunks aren't consistently placed
on servers based on their space partition, prohibiting some push-down
optimizations that rely on consistent partitioning.

This change ensures the ordinal of a space slice is calculated as if
all slices in the dimension are pre-existing. This might still lead to
inconsistencies during times of repartioning, but fixes issues that
occur initially when no slices exists.