Use a custom ServerScan node to scan servers
A new custom ServerScan node is now used to scan remote server
hypertables. Previously, a foreign scan was used for this purpose, but
this proved to have some issues that couldn't handle queries like,
e.g.:

```
SELECT device, avg(temp)
FROM hyper
WHERE time > '2019-06-01 01:00'
GROUP BY 1;
```

When scanning a foreign server this failed with an error saying that
"time" is not in the target list. The reason this happened is that we
treated the server scan as an upper relation, while it really is more
akin to a base relation. Upper relations invoke code with certain
expectations on target lists and restriction clauses (setrefs.c:
set_foreignscan_references), leading to the error.

The alternative is to treat the per-server scan as a base rel
(non-zero `scan_relid`), but this doesn't work with the ForeignScan
executor node because it expects any base rels to map to a "real"
foreign table in the database; thus the foreign scan fails with an
error when trying to lookup the foreign table.

A CustomScan implementation gives better control for handling query
optimzations and custom functionality that goes the beyond the
capabilities and expectations of ForeignScans. This also gives us a
better setup for future modifications to remote hypertable scans.