Support different clang-format names
On some distributions, the default version of clang-format is named
'clang-format' while other versions are available as 'clang-format-N',
where 'N' is the major version number of clang-format.

Current custom rules and scripts to format the entire repository expect
the name to be 'clang-format', which causes either lookup to fail or
execution to fail.

This commit changes the custom rule to look for clang-format-7 and
clang-format, in that order, before checking the version. In addition,
the scripts that execute clang-format are updated to support whatever
name the CMakeLists.txt found.