Re-configure cached remote connections
When a new connection is fetched from the cache, it can potentially
have broken since last use, so we extend the cache with a check
function that can be used to check if the cache entry is valid.

If the cache entry is not valid, it will be re-created instead of
updated.

If the cache entry is valid, or the cache entry check function is not
defined, the entry will be updated rather than re-created.

For the connection cache, a connection is invalid if it is still in
processing state. In that case, a previous use was aborted early and
the connection need to be re-created.