Block chunk creation by insert on data nodes
It is unsafe to insert data directly on data nodes if it would create
a new chunk, since chunk creation must be coordinated by the access
node. This change blocks such chunk creation but still allows
inserting tuples on data nodes that do not create new chunks and
otherwise adhere to existing constraints.