Don't fail `add_data_node()` when database exists
When the remote database exists `add_data_node()` currently fails
unless `if_not_exists` is true.

Typically, `IF NOT EXISTS` mean that an object creation should not
generate an error. Regardless of the `IF NOT EXISTS` the object is
still created.

The current behavior is not in line with what is expected so this
commit fixes this by not aborting the command if the remote database
exists and only print a notice that this step is skipped.

If the database exists, it is checked that the user have `CREATE`
privileges on the database before proceeding.