Add distributed_exec() function
This function allows users to execute a SQL query on a list of data nodes.
Original purpose of it is to provide a users a way to create roles on data nodes.

Current implementation is quite straightforward. Just execute any provided query
on a list of data nodes using current user id.
The function does not return or print any result values. In case of error, it will
print data node name and a related error message.