Remove cascade option from delete_data_node
The `cascade` option was added earlier since it was necessary to allow
cascading the delete of user mappings when removing the server objects.
Since the user mappings are removed from the code, the `cascade` option
is not needed any more.

This commit remove the option and fix all the tests.