Fix crash for add_data_node
It was possible to dereference a null pointer in
`add_data_node_internal` and it was not caught in tests because the
variable `server` was not initialized, resulting in a garbage value.
The variable was later used to retrieve a `Oid`, which happened to be
in valid memory and which was not used by the caller, so the faulty
reference was did not cause a crash.

Fixed the issue by reading the `Oid` from the proper place and also
initializing `server` to NULL so that is will crash if dereferenced.
Also removed the `server_id` variable that is not used.