Show NOTICE message on a distributed DROP DATABASE
This patch adds a way to check and print a notice message
to a user who wants to drop an access node database.

Since database drop can only be done using different database
connection, this functionality is implemented inside loader
extension.

Functionality of the security labels are used in order to mark
a distributed database and make this information accessible
by other databases in pg_shseclabel table.