Fix push down when hitting only one node
For some queries, with only one node "hit', push down can be forced by
matching the partitioning attributes with the group by expression.

However, push down didn't happen in some of these cases because of a
bug in `force_group_by_push_down` that didn't properly update the
number of attributes in the partitioning expression.

This change fixes the bug to always set the number of partitioning
attributes to match the group by clause.