Use PostgreSQL source include path when defined
This adds the include dir from the PostgreSQL source code to the
include path when PG_SOURCE_DIR is set (or detected) by CMake. This
allows the build to pick up (debug) settings in header files that were
configured when the PostgreSQL source was built. It also makes it
easier for IDEs to jump between the TimescaleDB source code and the
PostgreSQL source code.

This change also fixes a number of compiler/code warnings that were
discovered in the process.