Propagate grants when creating hypertables
When creating a hypertable, grants were not propagated to the table on
the remote node, which causes later statements to fail when not
executed as the owner of the table.

This commit deparse grant statements from the table definition and add
the grants to the deparsed statement to send when creating the table on
the data node.