Use qualified table name for chunks_in
If a table contains a column with a same name as table name then query parser
will get confused when parsing `chunks_in` function. The parser would think
that we are passing in column instead of a table. Using qualified table name
fixes this problem. Note that we needed to expand table using .* in order to
avoid parser confusion caused by schema.table syntax.