Fix crash when cancelling long running distributed insert
Problem lies in incorrect handling of reset callback for the
ChunkInsertState during the transaction abort procedure, which frees
parent memory context.

Because reset callbacks are executed only after all child memory context
got deleted, it is possible to end up in the sutiation when the context
is already deleted before this callback function being called.